Include waveAmpTimex for the waveguage in the etaImport class
This only creates a zero matrix for initialization purposes. This is because the Wavestar model includes the wavegauge input so it needs a vector to run. This value is 0 since the etaImport option only provides a point measurement at the body origin. Possible in future code releases to include two point measurements to provide upstream values.