Update wave gauge calculations
New condition used to check if a wave gauge has been set. the waveGaugeExist variable is completely removed.

wavegauge1loc now initialized to an empty array as: []

Comparing to NaN values proved more difficult so the empty array is used instead. The location is easily checked in the if statements using the isempty() function. wave gauge amplitude are still preallocated in memory if needed to save time.

Also removed the etaImport from the first case in waveElevationGrid() function. The last case specific to etaImport is never reached. If the last case if never used, it should be removed and etaImport restored to case 1.