Doc review (#446)
* github test

* tutorials

* Update wecSimInputFile.m

* Update wecSimInputFile.m

* Update wecSimInputFile.m

* Revert "Update wecSimInputFile.m"

This reverts commit 4b5c1614bc89efae517199eba6e9acae4d78ac9c.

* Revert "Update wecSimInputFile.m"

This reverts commit bee94a7dfb78b91c351b5509ed7d6997fdff6c3e.

* Revert "Update wecSimInputFile.m"

This reverts commit 77ef9e2be4f9d716c1266165bb55df074d64defe.

* Revert "tutorials"

This reverts commit b14731fbe7b8ae6d5b9db40155456e9a9000c490.

* Revert "github test"

This reverts commit b0799968b303651a5baef2ce975dc6a5b7c83e0e.

* Clean up wave class

Clean up wave class for project and issue  #419. Should speed pre-processing time by only calculating wave gauge amplitudes when necessary.

* Update waveClass.m

wave gauge location updates

* Update wave gauge calculations

New condition used to check if a wave gauge has been set. the waveGaugeExist variable is completely removed.

wavegauge1loc now initialized to an empty array as: []

Comparing to NaN values proved more difficult so the empty array is used instead. The location is easily checked in the if statements using the isempty() function. wave gauge amplitude are still preallocated in memory if needed to save time.

Also removed the etaImport from the first case in waveElevationGrid() function. The last case specific to etaImport is never reached. If the last case if never used, it should be removed and etaImport restored to case 1.

* Update waveClass.m

Add all necessary function arguments to allow for writing the correct Paraview files when the wave option 'etaImport' is used.

* Update waveClass.m

update default wave gauge location from [] to [NaN, NaN] so that the API better reflects its dimensions when a user does set it. Alter isempty() condition to isnan() for the amplitude calculation.

* Cleanup responseClass inputs

Add hspressure, wpressurenl and wpressurel into the bodyOutput structure so that its easier to pass into responseClass() initialization.

* Add wave gauges to responseClass

Lump wave type, elevation and gauge elevations into a waveOutput struct for easier passing into responseClass() initialization.

* end stop implementation

* Prefer Static Links to Download Directive

This commit switches from using the :download: directive to static
links for the webinar slides files. The value of this changes is that
the link paths will not be recalculated every time that the docs are
built and only a single copy of the files is stored (per branch).

* merging old gh-pages PR with new format

* fixed library masks

* remove mask comment

* save library as 2019a

* adding drag bod and hardstop doc rsts

* minor edit hard stop doc

* implementing suggestions from PR 430

* library saved to 2015b

* update to simulink figures. Zoomed in for easier viewing

* updating tutorials docs

* manual merge of new formatting

Co-authored-by: dforbush2 <dforbus@sandia.gov>
Co-authored-by: Mathew Topper <damm_horse@yahoo.co.uk>
Co-authored-by: Kelley Ruehl <kmruehl@sandia.gov>