Update Radiation_IRF.m to calculate infinite frequency added mass
Updated the Radiation_IRF.m to calculate the infinite frequency added mass from the impulse response function. In theory, a single frequency can be used with the impulse response function to calculate the infinite frequency added mass; however, with numerical calculations there is some variation so the mean across the entire range of frequencies calculated in the hydrodynamic input file.