Update wecSimInitialize.m
Added additional checks to ensure that simu.morisonElement is of the right dimension as the number of bodies. There will be issues with backwards compatibility since in the past the simu.morisonElement is initialized as a [1 x 1] vector. To resolve this issue in the near-term, if the simulation starts with a [1 x 1] vector equal to 0, then WEC-Sim assumes there are no active Morison Elements and simu.morisonElement is initialized as a [n x 1] vector of zeros where n is equal to the number of bodies.