Getting the gdf output right
When trying to use meshConvert.m to convert from NEMOH to WAMIT, the resulting gdf file produced a "end-of-file" error when read from WAMIT. NDEF, which should contain the number of panels actually contained the number of nodes (which was much higher). Looking at the code, I didn't really get the intention of the author and modified it so that the resulting gdf file consists of rows with 12 values each (4 nodes à 3 coordinates, forming one panel), as defined in the WAMIT v7.4 manual. The NDEF contains now the number of panels. All changes were made without considering any side effects on other conversion. Thus it would be great if someone with a deeper understanding could have a look on the proposed changes.