CI: Split docs jobs | Add color to docs logs | Cancel runs on new push | Add 2021b to MATLAB versions (#862)
* Split docs CI into test and build jobs

This PR splits the docs CI workflow into two independent jobs. The
first "test" job checks the current branch with any warnings
triggering a failure. The second jobs builds and deploys the
production docs, allowing any warnings to pass. This allows the
commit author to see any new issues they may be adding to the docs
while not stopping the docs being published.

* Try to get color output in log

* Remove redundant steps for branch test

* Fix duplicate targets using anonymous references

* Cancel previous runs if new commits are made

* Fix spelling mistake to test concurrency

* Limit concurrency to pull requests for the unit tests

This is to ensure all commits on the master and dev branches are
tested, which will be important if we are going to measure
coverage using an external service like codecov.

* Fix another spelling mistake

* Add R2021b to explicit MATLAB versions tested