Update radiationIRF.m (#1045)
This is a small fix to ensure that the infinite frequency added mass is calculated using the same range of frequencies used to calculate the impulse response function. Working with a user, we realized that the infinite frequency added mass calculation was flawed if there were low or high range of frequencies where the hydrodynamic coefficients were suspect. This arose in a Capytaine run where low frequency behavior impacted the calculation.