Create some scripts to run Nemoh, define num freqs and compare v2 to v3
It takes a really long time to run the default number of frequencies for each of these cases.

This run_nemoh.py script enables the user to specify the number of frequencies and the version of Nemoh (main(num_freqs=6, version='v2.3')). This makes it much faster to run a select number of frequencies for comparing the results between different Nemoh versions.